Change durationMax to float.
#17899:As a user, I can create activity definition at organization level

+++ ocp-fis/src/main/java/gov/samhsa/ocp/ocpfis/service/dto/TimingDto.java
@@ -12,6 +12,6 @@ import java.time.Duration;
 @AllArgsConstructor
 @NoArgsConstructor
 public class TimingDto {
-    private int durationMax;
+    private float durationMax;
     private int frequency;
 }
